HB-11: refactor
